monkey working

+++ OpenGLExercise/src/IndexBuffer.h
@@ -4,7 +4,8 @@
 class IndexBuffer
 {
 public:
-	IndexBuffer(const unsigned int* data, unsigned int count);
+	//IndexBuffer(const unsigned int* data, unsigned int count);
+	IndexBuffer(const unsigned short* data, unsigned int count);
 	~IndexBuffer();
 
 	void Bind() const;
